refactor(api): move controller and route
