--- api/lib/application/campaigns/index.js
@@ -1,29 +1,0 @@
-import Joi from 'joi';
-
-import { identifiersType } from '../../../src/shared/domain/types/identifiers-type.js';
-import { campaignController } from './campaign-controller.js';
-
-const register = async function (server) {
-  server.route([
-    {
-      method: 'GET',
-      path: '/api/campaigns/{id}/analyses',
-      config: {
-        validate: {
-          params: Joi.object({
-            id: identifiersType.campaignId,
-          }),
-        },
-        handler: campaignController.getAnalysis,
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            "- Récupération de l'analyse de la campagne par son id",
-        ],
-        tags: ['api', 'campaign'],
-      },
-    },
-  ]);
-};
-
-const name = 'campaigns-api';
-export { name, register };
