+++ api/lib/routes.js
@@ -3,7 +3,6 @@ import * as adminMembers from './application/admin-members/index.js';
 import * as authentication from './application/authentication/index.js';
 import * as cache from './application/cache/index.js';
 import * as campaignParticipations from './application/campaign-participations/index.js';
-import * as campaigns from './application/campaigns/index.js';
 import * as certificationCandidates from './application/certification-candidates/index.js';
 import * as certificationCenterInvitations from './application/certification-center-invitations/index.js';
 import * as certificationCenterMemberships from './application/certification-center-memberships/index.js';
@@ -42,7 +41,6 @@ const routes = [
   authentication,
   cache,
   campaignParticipations,
-  campaigns,
   certificationCandidates,
   certificationCenters,
   certificationCenterInvitations,
