+++ api/tests/prescription/campaign/unit/application/campaign-controller_test.js
@@ -1,8 +1,8 @@
-import { campaignController } from '../../../../lib/application/campaigns/campaign-controller.js';
-import { UserNotAuthorizedToAccessEntityError } from '../../../../lib/domain/errors.js';
-import { usecases } from '../../../../src/prescription/campaign/domain/usecases/index.js';
-import { LOCALE } from '../../../../src/shared/domain/constants.js';
-import { catchErr, expect, hFake, sinon } from '../../../test-helper.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../../lib/domain/errors.js';
+import { campaignController } from '../../../../../src/prescription/campaign/application/campaign-controller.js';
+import { usecases } from '../../../../../src/prescription/campaign/domain/usecases/index.js';
+import { LOCALE } from '../../../../../src/shared/domain/constants.js';
+import { catchErr, expect, hFake, sinon } from '../../../../test-helper.js';
 
 const { FRENCH_SPOKEN } = LOCALE;
