--- api/tests/unit/application/campaign/index_test.js
@@ -1,32 +1,0 @@
-import { campaignController } from '../../../../lib/application/campaigns/campaign-controller.js';
-import * as moduleUnderTest from '../../../../lib/application/campaigns/index.js';
-import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
-
-describe('Unit | Application | Router | campaign-router ', function () {
-  describe('GET /api/campaigns/{id}/analyses', function () {
-    it('should return 200', async function () {
-      // given
-      sinon.stub(campaignController, 'getAnalysis').callsFake((request, h) => h.response('ok').code(200));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/campaigns/1/analyses');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    it('should return 400', async function () {
-      // given
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/campaigns/wrong_id/analyses');
-
-      // then
-      expect(response.statusCode).to.equal(400);
-    });
-  });
-});
